Melakukan Filter

+++ backend/database/migrations/2026_08_29_110947_remove_detail_pinjam_id_from_pengembalian_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('pengembalian', function (Blueprint $table) {
-            $table->dropColumn('detail_pinjam_id');
-        });
+        if (Schema::hasColumn('pengembalian', 'detail_pinjam_id')) {
+            Schema::table('pengembalian', function (Blueprint $table) {
+                $table->dropColumn('detail_pinjam_id');
+            });
+        }
     }
 
     /**
